Replace share icons to separate file for .web and .native
