+++ src/components/invite/ShareIcons/ShareIcons.native.js
@@ -1,0 +1,1 @@
+export default () => null
