+++ src/components/invite/constants/constants.js
@@ -1,0 +1,6 @@
+export const shareTitle = 'I signed up to GoodDollar. Join me.'
+export const shareMessage =
+  'Hi,\nIf you believe, like me, in economic inclusion and the distribution of prosperity for all, then I invite you to sign up for GoodDollar, create your own basic income wallet and start collecting your daily digital income.\nUse my invite link and receive an extra 50G$ bonus\n\n'
+
+export const shortShareMessage =
+  'Hi,\nIf you believe in economic inclusion and distribution of prosperity for all, sign up for a GoodDollar wallet and start collecting daily digital income. Use my invite link and receive an extra 50G$\n\n'
