+++ src/components/invite/constants/index.js
@@ -1,0 +1,1 @@
+export { shareMessage, shortShareMessage, shareTitle } from './constants'
